split `blockchain/helper` functions
